rename package name crd to volumesnapshot
