+++ snapshot/pkg/controller/cache/desired_state_of_world.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/golang/glog"
 
-	crdv1 "github.com/kubernetes-incubator/external-storage/snapshot/pkg/apis/crd/v1"
+	crdv1 "github.com/kubernetes-incubator/external-storage/snapshot/pkg/apis/volumesnapshot/v1"
 )
 
 // DesiredStateOfWorld defines a set of thread-safe operations supported on
