+++ snapshot/pkg/volume/interface.go
@@ -19,7 +19,7 @@ package volume
 import (
 	"k8s.io/api/core/v1"
 
-	crdv1 "github.com/kubernetes-incubator/external-storage/snapshot/pkg/apis/crd/v1"
+	crdv1 "github.com/kubernetes-incubator/external-storage/snapshot/pkg/apis/volumesnapshot/v1"
 	"github.com/kubernetes-incubator/external-storage/snapshot/pkg/cloudprovider"
 )
